LayoutTests/editing/undomanager/document-has-undomanager.html

<!doctype html>
<html>
<head>
<title>document.undoManager</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    assert_own_property(document, "undoManager");
    assert_idl_attribute(document.undoManager, "undo");
    assert_idl_attribute(document.undoManager, "redo");
    assert_idl_attribute(document.undoManager, "clearUndo");
    assert_idl_attribute(document.undoManager, "clearRedo");
})
</script>
</body>
</html>
